probes: provide our own free function
liburing has a convenience function to get a io_uring_probe that can
be used to test for valid features. The function returns a probe that
the caller has to then deallocate.

That's not a problem if we are dealing with languanges like C and C++
but when trying to uring liburing bindings in languages with a more
well defined barrier into the C interface, calling free() can be at
best cumbersome and at worst just wrong as the language may be using
a different allocator under the hood.

Since we provide an allocated version of probe, provide users with
a way to deallocate it.